Be certified by System Tests
This is where the magic of containers come into play. I've written
a test suite that uses an actual, running application by using the
containers.

In the end, when you want to publish an image on docker hub, in
around a minute you will run the unit tests, build the image,
create docker containers in order to run system tests, tag and
publish the image.